+++ src/create/CreatePage.js
@@ -3,7 +3,7 @@ import styled from 'styled-components/macro'
 import PropTypes from 'prop-types'
 import Page from '../common/Page'
 import Header from '../common/Header'
-import Navigation from '../common/Navigation'
+import Navigation from '../app/Navigation'
 
 CreatePage.propTypes = {
   onSubmit: PropTypes.func
