--- src/notes/TagFilter.stories.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import TagFilter from './TagFilter'
-
-
-export default {
-  title: 'TagFilter'
-}
-
-export const tagFilter = () => (
-  <TagFilter>
-   
-  </TagFilter>
-)
