+++ src/stories/FilterButton.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
-import FilterButton from './FilterButton'
+import FilterButton from '../notes/FilterButton'
 
 export default {
   title: 'FilterButton'
