+++ src/stories/Tag.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Tag from './Tag'
+import Tag from '../notes/Tag'
 
 export default {
   title: 'Tag'
