+++ src/stories/TagFilter.stories.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import TagFilter from '../notes/TagFilter'
+
+export default {
+  title: 'TagFilter'
+}
+
+export const tagFilter = () => <TagFilter></TagFilter>
